Rename priority to score
