+++ src/optimizer/CMakeLists.txt
@@ -28,5 +28,4 @@ nebula_add_library(
     rule/PushFilterDownTagIndexScanRule.cpp
 )
 
-
 nebula_add_subdirectory(test)
